Move configuration to inner class.
